requirements.txt
certifi==2023.11.17 ; python_version >= "3.8" and python_version < "3.13"
charset-normalizer==3.3.2 ; python_version >= "3.8" and python_version < "3.13"
click==8.1.7 ; python_version >= "3.8" and python_version < "3.13"
colorama==0.4.6 ; python_version >= "3.8" and python_version < "3.13" and platform_system == "Windows"
dnspython==2.5.0 ; python_version >= "3.8" and python_version < "3.13"
docker==7.0.0 ; python_version >= "3.8" and python_version < "3.13"
idna==3.6 ; python_version >= "3.8" and python_version < "3.13"
markdown-it-py==3.0.0 ; python_version >= "3.8" and python_version < "3.13"
mdurl==0.1.2 ; python_version >= "3.8" and python_version < "3.13"
packaging==23.2 ; python_version >= "3.8" and python_version < "3.13"
pygments==2.17.2 ; python_version >= "3.8" and python_version < "3.13"
pymongo==4.6.1 ; python_version >= "3.8" and python_version < "3.13"
pywin32==306 ; python_version >= "3.8" and python_version < "3.13" and sys_platform == "win32"
requests==2.31.0 ; python_version >= "3.8" and python_version < "3.13"
rich==13.7.0 ; python_version >= "3.8" and python_version < "3.13"
ruamel-yaml-clib==0.2.8 ; platform_python_implementation == "CPython" and python_version < "3.13" and python_version >= "3.8"
ruamel-yaml==0.17.40 ; python_version >= "3.8" and python_version < "3.13"
shellingham==1.5.4 ; python_version >= "3.8" and python_version < "3.13"
typer[all]==0.9.0 ; python_version >= "3.8" and python_version < "3.13"
typing-extensions==4.9.0 ; python_version >= "3.8" and python_version < "3.13"
unique-names-generator==1.0.2 ; python_version >= "3.8" and python_version < "3.13"
urllib3==2.1.0 ; python_version >= "3.8" and python_version < "3.13"
